// src/enums/ETechnology.tsx
export enum ETechnology {
    Cpp = "cpp",
    CSharp = "c-sharp",
    Python = "python",
    React = "react",
    Fritzing = "fritzing",
    CSS = "CSS",
    HTML = "HTML",
    Firebase = "firebase",
    JavaScript = "javaScript",
    RaspberryPi = "raspberry-pi",
    OpenSCAD = "open-scad",
    MongoDB = "mongoDB",
    Twitter = "twitter",
    PostgresSQL = "postgres-sql",
    Arduino = "arduino",
    Instagram = "instagram",
    Facebook = "facebook",
    Youtube = "youtube",
    LinkedIn = "linkedin",
    Twitch = "twitch",
    TypeScript = "typeScript",
    Github = "github",
    Unity = "unity"
}